backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg[binary]==3.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.0.0
python-multipart==0.0.7
Pillow==11.1.0
